Use more specific name for testing containers
If test runs of multiple major versions are run on the same host, we can
end up unintentionally deleting containers while the tests are running.
By adding the version to the test container name we should be able to
avoid that. Requires changes in template generation, tests, and test
scripts.

Change-Id: I5f264ca06548020cc85be5b8055a04fa7f354dd5